+++ app/helpers/application_helper.rb
@@ -10,17 +10,9 @@ module ApplicationHelper
     end
   end
 
-  def month_timespan(stat)
-    month = DateTime.new(stat.year, stat.month).in_time_zone(Time.zone)
-    start_at = month.beginning_of_month.to_time.strftime('%Y-%m-%dT%H:%M')
-    end_at = month.end_of_month.to_time.strftime('%Y-%m-%dT%H:%M')
-
-    { start_at:, end_at: }
-  end
-
   def year_timespan(year)
-    start_at = Time.utc(year).in_time_zone('Europe/Berlin').beginning_of_year.strftime('%Y-%m-%dT%H:%M')
-    end_at = Time.utc(year).in_time_zone('Europe/Berlin').end_of_year.strftime('%Y-%m-%dT%H:%M')
+    start_at = Time.utc(year).in_time_zone(Time.zone).beginning_of_year.strftime('%Y-%m-%dT%H:%M')
+    end_at = Time.utc(year).in_time_zone(Time.zone).end_of_year.strftime('%Y-%m-%dT%H:%M')
 
     { start_at:, end_at: }
   end