+++ spec/factories/visits.rb
@@ -4,5 +4,10 @@ FactoryBot.define do
   factory :visit do
     area
     user
+    started_at { Time.zone.now }
+    ended_at { Time.zone.now + 1.hour }
+    duration { 1.hour }
+    name { 'Visit' }
+    status { 'pending' }
   end
 end