+++ spec/requests/visits_spec.rb
@@ -12,94 +12,84 @@ require 'rails_helper'
 # of tools you can use to make these specs even more expressive, but we're
 # sticking to rails and rspec-rails APIs to keep things simple and stable.
 
-RSpec.describe "/visits", type: :request do
-  
+RSpec.describe '/visits', type: :request do
   # This should return the minimal set of attributes required to create a valid
   # Visit. As you add validations to Visit, be sure to
   # adjust the attributes here as well.
-  let(:valid_attributes) {
-    skip("Add a hash of attributes valid for your model")
-  }
+  let(:valid_attributes) do
+    skip('Add a hash of attributes valid for your model')
+  end
 
-  let(:invalid_attributes) {
-    skip("Add a hash of attributes invalid for your model")
-  }
+  let(:invalid_attributes) do
+    skip('Add a hash of attributes invalid for your model')
+  end
 
-  describe "GET /index" do
-    it "renders a successful response" do
+  describe 'GET /index' do
+    it 'renders a successful response' do
       Visit.create! valid_attributes
       get visits_url
       expect(response).to be_successful
     end
   end
 
-  describe "GET /show" do
-    it "renders a successful response" do
+  describe 'GET /show' do
+    it 'renders a successful response' do
       visit = Visit.create! valid_attributes
       get visit_url(visit)
       expect(response).to be_successful
     end
   end
 
-  describe "GET /new" do
-    it "renders a successful response" do
-      get new_visit_url
-      expect(response).to be_successful
-    end
-  end
-
-  describe "GET /edit" do
-    it "renders a successful response" do
+  describe 'GET /edit' do
+    it 'renders a successful response' do
       visit = Visit.create! valid_attributes
       get edit_visit_url(visit)
       expect(response).to be_successful
     end
   end
 
-  describe "POST /create" do
-    context "with valid parameters" do
-      it "creates a new Visit" do
-        expect {
+  describe 'POST /create' do
+    context 'with valid parameters' do
+      it 'creates a new Visit' do
+        expect do
           post visits_url, params: { visit: valid_attributes }
-        }.to change(Visit, :count).by(1)
+        end.to change(Visit, :count).by(1)
       end
 
-      it "redirects to the created visit" do
+      it 'redirects to the created visit' do
         post visits_url, params: { visit: valid_attributes }
         expect(response).to redirect_to(visit_url(Visit.last))
       end
     end
 
-    context "with invalid parameters" do
-      it "does not create a new Visit" do
-        expect {
+    context 'with invalid parameters' do
+      it 'does not create a new Visit' do
+        expect do
           post visits_url, params: { visit: invalid_attributes }
-        }.to change(Visit, :count).by(0)
+        end.to change(Visit, :count).by(0)
       end
 
-    
       it "renders a response with 422 status (i.e. to display the 'new' template)" do
         post visits_url, params: { visit: invalid_attributes }
         expect(response).to have_http_status(:unprocessable_entity)
       end
-    
     end
   end
 
-  describe "PATCH /update" do
-    context "with valid parameters" do
-      let(:new_attributes) {
-        skip("Add a hash of attributes valid for your model")
-      }
+  describe 'PATCH /update' do
+    context 'with valid parameters' do
+      let(:new_attributes) do
+        skip('Add a hash of attributes valid for your model')
+      end
 
-      it "updates the requested visit" do
+      it 'updates the requested visit' do
         visit = Visit.create! valid_attributes
         patch visit_url(visit), params: { visit: new_attributes }
         visit.reload
-        skip("Add assertions for updated state")
+        skip('Add assertions for updated state')
       end
 
-      it "redirects to the visit" do
+      it 'redirects to the visit' do
         visit = Visit.create! valid_attributes
         patch visit_url(visit), params: { visit: new_attributes }
         visit.reload
@@ -107,26 +97,24 @@ RSpec.describe "/visits", type: :request do
       end
     end
 
-    context "with invalid parameters" do
-    
+    context 'with invalid parameters' do
       it "renders a response with 422 status (i.e. to display the 'edit' template)" do
         visit = Visit.create! valid_attributes
         patch visit_url(visit), params: { visit: invalid_attributes }
         expect(response).to have_http_status(:unprocessable_entity)
       end
-    
     end
   end
 
-  describe "DELETE /destroy" do
-    it "destroys the requested visit" do
+  describe 'DELETE /destroy' do
+    it 'destroys the requested visit' do
       visit = Visit.create! valid_attributes
-      expect {
+      expect do
         delete visit_url(visit)
-      }.to change(Visit, :count).by(-1)
+      end.to change(Visit, :count).by(-1)
     end
 
-    it "redirects to the visits list" do
+    it 'redirects to the visits list' do
       visit = Visit.create! valid_attributes
       delete visit_url(visit)
       expect(response).to redirect_to(visits_url)