+++ CHANGELOG.md
@@ -6,7 +6,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
 
-## [0.9.5] — 2024-07-22
+## [0.9.5] — 2024-07-26
 
 ### Added
 
@@ -14,9 +14,13 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - A background job to calculate your visits. This job will calculate your visits based on the areas you've created.
 - Visits page. This page will show you all your visits, calculated based on the areas you've created. You can see the date and time of the visit, the area you've visited, and the duration of the visit.
 - A possibility to confirm or decline a visit. When you create an area, the visit is not calculated immediately. You need to confirm or decline the visit. You can do this on the Visits page. Click on the visit, then click on the "Confirm" or "Decline" button. If you confirm the visit, it will be added to your timeline. If you decline the visit, it will be removed from your timeline.
-- [ ] Glue two consecutive visits if there are no points between them
-- [ ] Group visits by day and paginate them
+- [ ] Tests
+- [x] Swagger doc for Areas endpoint
+- [ ] Atomicity for visits creation
 
+### Fixed
+
+- A route popup now correctly shows distance made in the route, not the distance between first and last points in the route.
 ---
 
 ## [0.9.4] — 2024-07-21